Версия 1.2.5 Основное - проведена ревизия и значительные изменения кода в информационном обмене. Изменена структура БД. Требуется обновить версию
git-svn-id: svn+ssh://svn.code.sf.net/p/dynamicdocs/code@690 05742d63-fc59-4ccd-b3f0-d7e2a2a8a88d

+++ ddocserver/internals/functions/inf_exchange/addsyncrecord.sql
@@ -1,4 +1,4 @@
-create or replace function addSyncRecord(int4, int4, varchar, varchar, varchar, int4, int4) returns int4 as
+create or replace function addSyncRecord(int4, int8, varchar, varchar, varchar, int4, int4) returns int8 as
 $BODY$
 declare
     idOrganization alias for $1;
@@ -20,7 +20,7 @@ declare
                               -- 12 - files (transferred by parts) (creation, updating)
                               -- 13 - files (transferred by parts) (deleting)
 
-    idRecord int4;
+    idRecord int8;
 begin
 
 
@@ -60,7 +60,7 @@ create or replace function syncRecords() returns trigger as
 $BODY$
 declare
     theTable varchar;
-    idEntity int4;
+    idEntity int8;
     entityUID varchar;
     entityIOUID varchar;
     syncType int4;
